Use CachedProp constructor
