+++ packages/react-native-nitro-modules/cpp/templates/IsFunctionProp.hpp
@@ -1,0 +1,25 @@
+//
+//  IsFunctionProp.hpp
+//  NitroModules
+//
+//  Created by Marc Rousavy on 21.06.24.
+//
+
+#pragma once
+
+#include <type_traits>
+
+namespace margelo::nitro {
+
+using namespace facebook;
+
+template <typename T>
+struct IsFunctionProp : std::false_type {};
+
+template <typename TResult, typename... TArgs>
+struct IsFunctionProp<std::function<TResult(TArgs...)>> : std::true_type {};
+
+template <typename T>
+struct IsFunctionProp<std::optional<T>> : IsFunctionProp<T> {};
+
+} // namespace margelo::nitro
